# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "grtblog 后端接口",
        "title": "grtblog API v2",
        "contact": {},
        "version": "2.0.0"
    },
    "basePath": "/api/v2",
    "paths": {
...
        "/thinkings": {
            "get": {
                "tags": [
                    "Thinking"
                ],
                "summary": "获取思考列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页大小",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.ListThinkingResp"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "JWTAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Thinking"
                ],
                "summary": "发布思考",
                "parameters": [
                    {
                        "description": "创建参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/contract.CreateThinkingReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.ThinkingResp"
                        }
                    }
                }
            }
        },
        "/thinkings/{id}": {
            "get": {
                "tags": [
                    "Thinking"
                ],
                "summary": "获取单个思考",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "思考ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.ThinkingResp"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "JWTAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Thinking"
                ],
                "summary": "更新思考",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "思考ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/contract.UpdateThinkingReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "JWTAuth": []
                    }
                ],
                "tags": [
                    "Thinking"
                ],
                "summary": "删除思考",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "思考ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "contract.CreateThinkingReq": {
            "type": "object",
            "required": [
                "content"
            ],
            "properties": {
                "allowComment": {
                    "type": "boolean"
                },
                "content": {
                    "type": "string"
                }
            }
        },
...
        "contract.ListThinkingResp": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/contract.ThinkingResp"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "contract.ThinkingMetrics": {
            "type": "object",
            "properties": {
                "comments": {
                    "type": "integer"
                },
                "likes": {
                    "type": "integer"
                },
                "views": {
                    "type": "integer"
                }
            }
        },
        "contract.ThinkingResp": {
            "type": "object",
            "properties": {
                "allowComment": {
                    "type": "boolean"
                },
                "authorId": {
                    "type": "integer"
                },
                "authorName": {
                    "type": "string"
                },
                "avatar": {
                    "type": "string"
                },
                "commentId": {
                    "type": "integer"
                },
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isHot": {
                    "type": "boolean"
                },
                "metrics": {
                    "$ref": "#/definitions/contract.ThinkingMetrics"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "contract.UpdateThinkingReq": {
            "type": "object",
            "required": [
                "content"
            ],
            "properties": {
                "allowComment": {
                    "type": "boolean"
                },
                "content": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}